Patch @metamask/scure-bip39 to reset encodedMnemonicUint8Array
Without this change, when using a software like ProcessHacker2 on
Windows to inspect the processes, one can find the mnemonic as a string
every time the wallet unlocks. This commit resets the point in memory
where the mnemonic as an Uint8Array is stored before it gets converted
to a seed.